Move version JSON code into website.py
This is a prerequisite to fixing the download jar option.
