+++ burger/website.py
@@ -21,8 +21,22 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 """
-import urllib
+import six.moves.urllib.request
 
+try:
+    import json
+except ImportError:
+    import simplejson as json
+
+VERSION_MANIFEST = "https://launchermeta.mojang.com/mc/game/version_manifest.json"
+LEGACY_VERSION_META = "https://s3.amazonaws.com/Minecraft.Download/versions/%(version)s/%(version)s.json" # DEPRECATED
+
+def _load_json(url):
+    stream = six.moves.urllib.request.urlopen(url)
+    try:
+        return json.load(stream)
+    finally:
+        stream.close()
 
 class Website(object):
     def __init__(self, username, password, version=999999):
@@ -31,8 +45,39 @@ class Website(object):
         self.version = version
 
     @staticmethod
+    def get_version_meta(version, verbose):
+        """
+        Gets a version JSON file, first attempting the to use the version manifest
+        and then falling back to the legacy site if that fails.
+        Note that the main manifest should include all versions as of august 2018.
+        """
+        version_manifest = _load_json(VERSION_MANIFEST)
+        for version_info in version_manifest["versions"]:
+            if version_info["id"] == version:
+                address = version_info["url"]
+                break
+        else:
+            if verbose:
+                print("Failed to find %s in the main version manifest; using legacy site" % version)
+                address = LEGACY_VERSION_META % {'version': version}
+        if verbose:
+            print("Loading version manifest for %s from %s" % (version, address))
+        return _load_json(address)
+
+    @staticmethod
+    def get_asset_index(version_meta, verbose):
+        """Downloads the Minecraft asset index"""
+        if "assetIndex" not in version_meta:
+            raise Exception("No asset index defined in the version meta")
+        asset_index = version_meta["assetIndex"]
+        if verbose:
+            print("Assets: id %(id)s, url %(url)s" % asset_index)
+        return _load_json(asset_index["url"])
+
+
+    @staticmethod
     def client_jar(path=None, reporthook=None, version="1.9"):
         url = "http://s3.amazonaws.com/Minecraft.Download/versions/%s/%s.jar" % (version, version)
         #url = "http://s3.amazonaws.com/MinecraftDownload/minecraft.jar" # 1.5.2
-        r = urllib.urlretrieve(url, filename=path, reporthook=reporthook)
+        r = six.moves.urllib.request.urlretrieve(url, filename=path, reporthook=reporthook)
         return r[0]
